Move tests for moved logic to new location
