+++ packages/framework/tests/Feature/RouteTest.php
@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Testing\Feature;
 
-use Hyde\Framework\Exceptions\RouteNotFoundException;
 use Hyde\Hyde;
-use Hyde\Pages\BladePage;
 use Hyde\Pages\MarkdownPage;
 use Hyde\Support\Facades\Render;
 use Hyde\Support\Models\Route;
@@ -67,39 +65,6 @@ class RouteTest extends TestCase
         $this->assertEquals($page->getOutputPath(), $route->getOutputPath());
     }
 
-    public function test_get_is_alias_for_get_from_key()
-    {
-        $this->assertEquals(\Hyde\Facades\Route::get('index'), \Hyde\Facades\Route::get('index'));
-    }
-
-    public function test_get_or_fail_throws_exception_if_route_is_not_found()
-    {
-        $this->expectException(RouteNotFoundException::class);
-        \Hyde\Facades\Route::getOrFail('not-found');
-    }
-
-    public function test_get_from_key_returns_route_from_router_index()
-    {
-        $this->assertEquals(new Route(BladePage::parse('index')), \Hyde\Facades\Route::get('index'));
-        $this->assertInstanceOf(Route::class, \Hyde\Facades\Route::get('index'));
-    }
-
-    public function test_get_from_returns_null_if_route_is_not_found()
-    {
-        $this->assertNull(\Hyde\Facades\Route::get('not-found'));
-    }
-
-    public function test_get_supports_dot_notation()
-    {
-        $this->file('_posts/foo.md');
-        $this->assertSame(\Hyde\Facades\Route::get('posts/foo'), \Hyde\Facades\Route::get('posts.foo'));
-    }
-
-    public function test_route_facade_all_method_returns_all_routes()
-    {
-        $this->assertEquals(Hyde::routes(), \Hyde\Facades\Route::all());
-    }
-
     public function test_get_link_returns_correct_path_for_root_pages()
     {
         $route = new Route(new MarkdownPage('foo'));
@@ -136,18 +101,6 @@ class RouteTest extends TestCase
         $this->assertEquals($route->getLink(), (string) $route);
     }
 
-    public function test_current_returns_current_route()
-    {
-        $route = new Route(new MarkdownPage('foo'));
-        Render::share('currentRoute', $route);
-        $this->assertEquals($route, \Hyde\Facades\Route::current());
-    }
-
-    public function test_current_returns_null_if_route_is_not_found()
-    {
-        $this->assertNull(\Hyde\Facades\Route::current());
-    }
-
     public function test_to_array_method()
     {
         $this->assertEquals([
@@ -182,11 +135,5 @@ class RouteTest extends TestCase
         $route = new Route(new MarkdownPage('foo'));
         $this->assertTrue($route->is(new RouteKey('foo')));
         $this->assertFalse($route->is(new RouteKey('bar')));
-    }
-
-    public function testExists()
-    {
-        $this->assertTrue(\Hyde\Facades\Route::exists('index'));
-        $this->assertFalse(\Hyde\Facades\Route::exists('not-found'));
     }
 }
