+++ packages/framework/tests/Unit/Facades/RouteFacadeTest.php
@@ -4,6 +4,12 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Testing\Unit\Facades;
 
+use Hyde\Facades\Route;
+use Hyde\Framework\Exceptions\RouteNotFoundException;
+use Hyde\Hyde;
+use Hyde\Pages\BladePage;
+use Hyde\Pages\MarkdownPage;
+use Hyde\Support\Facades\Render;
 use Hyde\Support\Models\Route as RouteModel;
 use Hyde\Testing\TestCase;
 
@@ -12,33 +18,54 @@ use Hyde\Testing\TestCase;
  */
 class RouteFacadeTest extends TestCase
 {
-    public function testFacadeMethodGet()
+    public function test_route_facade_all_method_returns_all_routes()
     {
-        $this->markTestIncomplete('TODO: Reimplement the tests');
-        // $this->assertSame(\Hyde\Facades\Route::get('index'), RouteModel::get('index'));
+        $this->assertEquals(Hyde::routes(), Route::all());
     }
 
-    public function testFacadeMethodGetOrFail()
+    public function test_get_is_alias_for_get_from_key()
     {
-        $this->markTestIncomplete('TODO: Reimplement the tests');
-        // $this->assertSame(\Hyde\Facades\Route::getOrFail('index'), RouteModel::getOrFail('index'));
+        $this->assertEquals(Route::get('index'), Route::get('index'));
     }
 
-    public function testFacadeMethodAll()
+    public function test_get_or_fail_throws_exception_if_route_is_not_found()
     {
-        $this->markTestIncomplete('TODO: Reimplement the tests');
-        // $this->assertSame(\Hyde\Facades\Route::all(), RouteModel::all());
+        $this->expectException(RouteNotFoundException::class);
+        Route::getOrFail('not-found');
     }
 
-    public function testFacadeMethodCurrent()
+    public function test_get_from_key_returns_route_from_router_index()
     {
-        $this->markTestIncomplete('TODO: Reimplement the tests');
-        // $this->assertSame(\Hyde\Facades\Route::current(), RouteModel::current());
+        $this->assertEquals(new RouteModel(BladePage::parse('index')), Route::get('index'));
+        $this->assertInstanceOf(RouteModel::class, Route::get('index'));
     }
 
-    public function testFacadeMethodExists()
+    public function test_get_from_returns_null_if_route_is_not_found()
     {
-        $this->markTestIncomplete('TODO: Reimplement the tests');
-        // $this->assertSame(\Hyde\Facades\Route::exists('index'), RouteModel::exists('index'));
+        $this->assertNull(Route::get('not-found'));
+    }
+
+    public function test_get_supports_dot_notation()
+    {
+        $this->file('_posts/foo.md');
+        $this->assertSame(Route::get('posts/foo'), Route::get('posts.foo'));
+    }
+
+    public function test_current_returns_current_route()
+    {
+        $route = new RouteModel(new MarkdownPage('foo'));
+        Render::share('currentRoute', $route);
+        $this->assertEquals($route, Route::current());
+    }
+
+    public function test_current_returns_null_if_route_is_not_found()
+    {
+        $this->assertNull(Route::current());
+    }
+
+    public function testExists()
+    {
+        $this->assertTrue(Route::exists('index'));
+        $this->assertFalse(Route::exists('not-found'));
     }
 }
